+++ packages/natives/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+
+- Added `hasConflicts()` method to detect unresolved merge conflicts in parsed files
+- Added `conflictCount()` method to count unresolved merge conflicts in the chunk tree
 
 ## [14.0.2] - 2026-04-09